+++ src/main/java/de/lemo/apps/restws/proxies/questions/QCourseActivity.java
@@ -27,6 +27,6 @@ public interface QCourseActivity {
 			@FormParam(MetaParam.USER_IDS) List<Long> users,
 			@FormParam(MetaParam.START_TIME) Long starttime,
 			@FormParam(MetaParam.END_TIME) Long endtime,
-			@FormParam(MetaParam.RESOLUTION) Integer resolution,
+			@FormParam(MetaParam.RESOLUTION) Long resolution,
 			@FormParam(MetaParam.TYPES) List<String> resourceTypes);
 }
